add pop_count

+++ poly_ring_arithmetic/src/poly_r_8192.h
@@ -35,7 +35,7 @@ extern void div_poly(poly *quo, poly *re, const poly *dividend, const poly *divi
 
 extern unsigned long get_deg(const poly *p);
 
-extern unsigned long long  pop_count(const poly *p2);
+extern int  pop_count(const poly *p2);
